MNIST Example added
