+++ src/L423/tools/visualize.py
@@ -1,0 +1,29 @@
+from matplotlib import pyplot as plt
+
+
+def show_image(image,normal=False):
+    plt.axis("off")
+    if(normal):
+        plt.imshow(image,cmap='gray',vmin=0,vmax=255)
+    else:
+        plt.imshow(image,cmap='gray')
+    plt.show()
+
+def show_filters(weight):
+    fig,axes = plt.subplots(1,weight.shape[0])
+    fig.set_size_inches(5*weight.shape[0], 5)
+    # fig.suptitle(f'plots of synaptic share weights for d = {weight.shape[0]}')
+    for i in range(weight.shape[0]):
+        axes[i].imshow(weight[i][0],cmap='gray')
+        axes[i].axis('off')
+    plt.show()
+        
+        
+def show_images(imgs,title,count):
+    fig,axes = plt.subplots(1,count)
+    fig.set_size_inches(5*count, 5)
+    plt.text(x=0.5, y=0.94, s=title, fontsize=28, ha="center", transform=fig.transFigure)
+    for i in range(count):
+        axes[i].imshow(imgs[i][0][0],cmap='gray')
+        axes[i].axis('off')
+    plt.show()
